add new test bag: phidgets_imu_upside_down
This is from a phidgets imu, running the phidgets_imu driver mounted on
a wheeled robot that drives a small trip through an office. The IMU is
mounted upside down (i.e., the chips on the board are on the upper side,
lin_acc.z is ~ -9.81). The USB port points to the right side of the
robot. Rotation quaternion from /base_link (x pointing forward, z
pointing up) to /imu is (0, 1, 0, 0).

Contains /imu/data_raw and /imu/mag topics.