Move rviz_imu_plugin headers to src
This is required so that the CMAKE_AUTOMOC line introduced in the
previous commit works.

Without this change,
imu_tools/rviz_imu_plugin/CMakeFiles/rviz_imu_plugin.dir/include/rviz_imu_plugin/moc_imu_display.cxx.o
would not be included in the library (it was in MOC_FILES before). This
led to an error when attempting to load the plugin in RViz.