Add option to display orientation in world frame (#69)
Per REP 145 IMU orientation is in the world frame. Rotating the
orientation data to transform into the sensor frame results in strange
behavior, such as double-rotation of orientation on a robot. Provide an
option to display orientation in the world frame, and enable it by
default. Continue to translate the position of the data to the sensor
frame.